--- src/components/mypage/image-upload/ImageUpload.module.scss
@@ -1,23 +1,0 @@
-@import '../../../styles/global.scss';
-
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 18.2rem;
-  height: 18.2rem;
-  background-color: #f5f5f5;
-  border-radius: 0.6rem;
-
-  @include mobile {
-    width: 10rem;
-    height: 10rem;
-  }
-}
-
-.img {
-  @include mobile {
-    width: 2rem;
-    height: 2rem;
-  }
-}
