+++ src/components/mypage/profile/Profile.module.scss
@@ -28,7 +28,7 @@
   }
 }
 
-.items {
+.form {
   display: flex;
   margin-top: 3.2rem;
   margin-bottom: 2rem;
@@ -43,17 +43,13 @@
 }
 
 .img-container {
-  position: relative;
-  overflow: hidden;
   display: flex;
   justify-content: center;
   align-items: center;
   width: 18.2rem;
-  height: 18.2rem;
-  border-radius: 0.6rem;
+  height: 20rem;
   background-color: #f5f5f5;
-
-  cursor: pointer;
+  border-radius: 0.6rem;
 
   @include mobile {
     width: 10rem;
@@ -61,11 +57,18 @@
   }
 }
 
+.img {
+  @include mobile {
+    width: 2rem;
+    height: 2rem;
+  }
+}
+
 .input {
   display: none;
 }
 
-.form {
+.items {
   display: flex;
   flex-direction: column;
   gap: 2rem;
